Address review: clamp on existing batch_size, drop transform_batch_size knob
Only raise datafusion.execution.batch_size when the read batch_size exceeds
DataFusion's default; never lower it. Removes the separate transform_batch_size
parameter per review feedback.
